test(cms): isolate sync apply coverage
